requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cohere==4.1.7
faiss-cpu==1.7.3
Flask==2.2.3
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
pandas==2.0.0
praw==7.7.0
prawcore==2.3.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
six==1.16.0
tzdata==2023.3
update-checker==0.18.0
urllib3==1.26.15
websocket-client==1.5.1
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
